+++ jum.py
@@ -5,7 +5,7 @@ from sys import exit as sys_exit
 from termios import tcflush, TCIOFLUSH
 from time import sleep
 from select import select
-from sys import stdin, stdout
+from sys import stdin
 
 import assets.gfx as gfx
 import assets.window as window
@@ -13,11 +13,6 @@ import assets.window as window
 frame_break = float(0.08) # Time to render a single frame.
 
 # Frames, logic of rendering.
-def flush_previous_frame(): # To render the game in a single frame.
-	for i in range(gfx.y_size - 1):
-		gfx.clearline()
-		stdout.flush()
-
 def frame():
 	global frame_break
 
@@ -28,10 +23,10 @@ def frame():
 		if random_delay == 0 and frame_break > 0.04:
 			frame_break -= 0.02 # Increase speed of the game.
 
-		flush_previous_frame()
+		gfx.flush_previous_frame()
 
 		window.score += 1 # Python has got infinite number precision.
-		window.enemy_x = randint(5, (gfx.x_size - window.player_x - 4))
+		window.enemy_x = randint(10, (gfx.x_size - window.player_x - 4))
 
 		window.idle_no_enemy()
 		sleep(randint(0, 1))
@@ -44,22 +39,22 @@ def keypress(): # Keyboard-event.
 
 
 window.idle_no_enemy()	# Initial frame.
-flush_previous_frame()
+gfx.flush_previous_frame()
 
 while 1: # Main game loop.
 	if keypress() == 1:
 		tcflush(stdin, TCIOFLUSH) # Flush input buffer.
 
 		for i in range(2): # Jump width. Must be smaller than player_x.
-			flush_previous_frame()
+			gfx.flush_previous_frame()
 			window.enemy_x -= 1 # Change enemy position.
 
 			if window.jump(window.enemy_x) == 0: # Enemy behind the player.
-				window.enemy_x = randint(5, (gfx.x_size - window.player_x - 4))
+				window.enemy_x = randint(10, (gfx.x_size - window.player_x - 4))
 
 			sleep(frame_break)
 
 	else:
-		flush_previous_frame()
+		gfx.flush_previous_frame()
 		frame()