+++ assets/window.py
@@ -10,21 +10,21 @@ if gfx.x_size % 2 == 0: # Player centering.
 else:
 	player_x = int((gfx.x_size + 1) / 2)
 
-enemy_x = randint(10, (gfx.x_size - player_x - 4))
+enemy_x = randint(10, (gfx.x_size - player_x - 2))
 after_enemy_x = gfx.y_size - enemy_x
 score = int(0)
 
 
 # Main areas to render on the screen.
 def environment(): # Non-playable area above the playable area.
-	for i in range(gfx.x_size):
+	for i in range(gfx.x_size): # Debug line.
 		print("#", end = "")
 
 	for i in range((gfx.y_size - 4) * gfx.x_size): # Render environment area.
 		print(end = " ") # Environment empty area..
 
 def empty(): # Upper part of the playable_area.
-	for i in range(gfx.x_size): # Spaces after left border.
+	for i in range(gfx.x_size):
 		print(end = " ")
 
 def render_player(): # Shows the player with specified position.
@@ -35,43 +35,30 @@ def render_player(): # Shows the player with specified position.
 
 def floor(): # Long block under the player and enemies.
 	score_len = len(str(score))
-	print(gfx.blue + "  Score:", score, gfx.default, end = "")
+	print(gfx.blue + " Score:", score, gfx.default, end = "")
 
-	for i in range(gfx.x_size - score_len - 10):
+	for i in range(gfx.x_size - score_len - 9):
 		print(gfx.blue + "#" + gfx.default, end = "")
 
 
-# Scenarios:
+
+# Scenarios of rendering:
 def jump(enemy_x): # Frame rendered when the player jumps.
 	environment()
-
-	# Upper part of the playable_area.
-	print("#", end = "") # Left border.
-
-	for i in range(player_x): # Spaces after left border.
+	render_player()
+	for i in range(gfx.x_size - player_x - 1): # Spaces before the left border.
 		print(end = " ")
 
-	gfx.player()
-
-	for i in range(gfx.x_size - player_x - 3): # Spaces before the left border.
-		print(end = " ")
-
-	print("#") # Right border.
-
-	# Lower part of the playable_area.
-	# Left, fixed border before the Player.
-	print("#", end = "")
-
-	for i in range(player_x + 1 + enemy_x):
+	# Lower area:
+	for i in range(player_x + enemy_x):
 		print(end = " ") # Spaces before the enemy, when the player jumps.
 
 	gfx.enemy()
 
-	for i in range(gfx.x_size - player_x - enemy_x - 4):
+	for i in range(gfx.x_size - player_x - enemy_x - 2):
 		print(end = " ") # Space chars after the Enemy.
 
-	print("#") # Border after the enemy.
-
+	print("") # Placeholder to put the keypress catcher to correct place.
 	floor()
 
 	if enemy_x + player_x < 0: # Enemy at the end of the map (left).
@@ -81,26 +68,22 @@ def idle(enemy_x): # Scenario when there are no activities from the player.
 	environment()
 	empty()
 
-	print("#", end = "")
-
+	# Lower area:
 	render_player()
-
 	for i in range(enemy_x): # Spaces before the enemy.
 		print(end = " ")
 
 	gfx.enemy()
 
-	for i in range(gfx.x_size - player_x - enemy_x - 4):
+	for i in range(gfx.x_size - player_x - enemy_x - 2):
 		print(end = " ") # Space chars after the Enemy.
 
-	print("#") # Border after the enemy.
-
+	print("") # Placeholder to put the keypress catcher to correct place.
 	floor()
 
 	if enemy_x == 0: # If colission then end the game.
 		score_len = len(str(score))
 		gfx.clearline()
-
 		print(gfx.blue + "Score:", score, gfx.default, end = "")
 
 		for i in range(gfx.x_size - score_len - 17):
@@ -111,19 +94,4 @@ def idle(enemy_x): # Scenario when there are no activities from the player.
 
 	if enemy_x < 0:
 		return 0
-
-def idle_no_enemy(): # Rendered like above but without enemy.
-	environment()
-	empty()
-
-	print("#", end = "")
-
-	render_player()
-
-	for i in range(gfx.x_size - player_x - 3):
-		print(end = " ")
-
-	print("#")
-
-	floor()