+++ assets/gfx.py
@@ -7,7 +7,7 @@ y_size, x_size = popen("stty size", "r").read().split()
 x_size = int(x_size)
 y_size = int(y_size)
 
-if x_size < 30 or y_size < 5:
+if x_size < 40 or y_size < 5:
 	print("To small window!")
 	sys_exit(1)