+++ docs/CHANGELOG.md
@@ -1,9 +1,10 @@
 ### jumpy-dot changelog
 
-#### wip v1.2
+#### wip v1.2 - what's new?
 Functions in window.py slightly straightened.
 Winsize check deleted to provide better accesbility.
-Real fullscreen rendering added.
+Real fullscreen rendering.
+Enemy position randomization.
 
 #### v1.1
 Speed up delay is now random (20%).